+++ engine/apps/api/tests/test_oncall_shift.py
@@ -8,7 +8,7 @@ from rest_framework.response import Response
 from rest_framework.test import APIClient
 
 from apps.api.permissions import LegacyAccessControlRole
-from apps.schedules.models import CustomOnCallShift, OnCallSchedule, OnCallScheduleWeb
+from apps.schedules.models import CustomOnCallShift, OnCallSchedule, OnCallScheduleCalendar, OnCallScheduleWeb
 
 
 @pytest.fixture()
@@ -57,6 +57,46 @@ def test_create_on_call_shift_rotation(on_call_shift_internal_api_setup, make_us
     assert response.status_code == status.HTTP_201_CREATED
     assert response.json() == expected_payload
     assert mock_refresh_schedule.called
+
+
+@pytest.mark.django_db
+def test_create_on_call_shift_rotation_invalid_type(
+    make_organization_and_user_with_plugin_token,
+    make_schedule,
+    make_user_auth_headers,
+):
+    organization, user, token = make_organization_and_user_with_plugin_token()
+    schedule = make_schedule(organization, schedule_class=OnCallScheduleCalendar)
+
+    client = APIClient()
+    url = reverse("api-internal:oncall_shifts-list")
+    start_date = timezone.now().replace(microsecond=0, tzinfo=None)
+
+    data = {
+        "name": "Test Shift",
+        "type": CustomOnCallShift.TYPE_ROLLING_USERS_EVENT,
+        "schedule": schedule.public_primary_key,
+        "priority_level": 1,
+        "shift_start": start_date.strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "shift_end": (start_date + timezone.timedelta(hours=2)).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "rotation_start": start_date.strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "until": None,
+        "frequency": 1,
+        "interval": 1,
+        "by_day": [
+            CustomOnCallShift.ICAL_WEEKDAY_MAP[CustomOnCallShift.MONDAY],
+            CustomOnCallShift.ICAL_WEEKDAY_MAP[CustomOnCallShift.FRIDAY],
+        ],
+        "week_start": CustomOnCallShift.ICAL_WEEKDAY_MAP[CustomOnCallShift.MONDAY],
+        "rolling_users": [[user.public_primary_key]],
+    }
+
+    with patch("apps.schedules.models.CustomOnCallShift.refresh_schedule") as mock_refresh_schedule:
+        response = client.post(url, data, format="json", **make_user_auth_headers(user, token))
+
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
+    assert response.data["type"][0] == "Invalid event type"
+    assert not mock_refresh_schedule.called
 
 
 @pytest.mark.django_db
@@ -1555,6 +1595,42 @@ def test_on_call_shift_preview(
         if not e["is_override"] and not e["is_gap"]
     ]
     assert returned_events == expected_events
+
+
+@pytest.mark.django_db
+def test_on_call_shift_preview_invalid_type(
+    make_organization_and_user_with_plugin_token,
+    make_user_auth_headers,
+    make_schedule,
+):
+    organization, user, token = make_organization_and_user_with_plugin_token()
+    client = APIClient()
+
+    schedule = make_schedule(organization, schedule_class=OnCallScheduleCalendar)
+
+    now = timezone.now().replace(hour=0, minute=0, second=0, microsecond=0)
+    start_date = now - timezone.timedelta(days=7)
+    request_date = start_date
+
+    url = "{}?date={}&days={}".format(
+        reverse("api-internal:oncall_shifts-preview"), request_date.strftime("%Y-%m-%d"), 1
+    )
+    shift_start = (start_date + timezone.timedelta(hours=12)).strftime("%Y-%m-%dT%H:%M:%SZ")
+    shift_end = (start_date + timezone.timedelta(hours=13)).strftime("%Y-%m-%dT%H:%M:%SZ")
+    shift_data = {
+        "schedule": schedule.public_primary_key,
+        "type": CustomOnCallShift.TYPE_ROLLING_USERS_EVENT,
+        "rotation_start": shift_start,
+        "shift_start": shift_start,
+        "shift_end": shift_end,
+        "rolling_users": [[user.public_primary_key]],
+        "priority_level": 2,
+        "frequency": CustomOnCallShift.FREQUENCY_DAILY,
+        "interval": 1,
+    }
+    response = client.post(url, shift_data, format="json", **make_user_auth_headers(user, token))
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
+    assert response.data["type"][0] == "Invalid event type"
 
 
 @pytest.mark.django_db